requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
dagster
dagit
pandas
python-dotenv
psycopg2
# For dbt transformation
# You may also install dbt-postgres via pip
ultralytics